Dev/isabella mendes

+++ app/src/main/java/com/example/library/Convidado.kt
@@ -4,6 +4,6 @@ data class Convidado(
     val id: Long,
     val nome: String,
     val descricao: String,
-    val evento_id: Int
+    val evento_id: Long
 )
